move /lib/tags -> /lib/tags-images
