+++ lib/admin/admin-tags-form/view.js
@@ -1,6 +1,6 @@
 import t from 't-component'
 import FormView from '../../form-view/form-view'
-import images from '../../tags/images'
+import images from '../../tags-images'
 import template from './template.jade'
 
 export default class TagForm extends FormView {
