+++ lib/models/tag.js
@@ -11,7 +11,7 @@ require('node-path')(module)
 
 var mongoose = require('mongoose')
 var Schema = mongoose.Schema
-var tagImages = require('lib/tags/images.js')
+var tagImages = require('lib/tags-images')
 var images = Object.keys(tagImages)
 var regex = require('lib/regexps')
