+++ lib/render/render.js
@@ -1,9 +1,9 @@
 import domify from 'domify'
 import merge from 'merge'
-import user from '../user/user.js'
 import t from 't-component'
-import tagImages from '../tags/images.js'
+import user from '../user/user'
 import config from '../config/config'
+import tagImages from '../tags-images'
 
 /**
  * Render default modules
