+++ lib/tags-images/index.js
@@ -5,38 +5,38 @@
 module.exports = {
   'arts': {
     name: 'arts',
-    url: '/tags/arts.svg'
+    url: '/tags-images/arts.svg'
   },
   'scale': {
     name: 'scale',
-    url: '/tags/scale.svg'
+    url: '/tags-images/scale.svg'
   },
   'police': {
     name: 'police',
-    url: '/tags/police.svg'
+    url: '/tags-images/police.svg'
   },
   'signs': {
     name: 'signs',
-    url: '/tags/signs.svg'
+    url: '/tags-images/signs.svg'
   },
   'transportation': {
     name: 'transportation',
-    url: '/tags/transportation.svg'
+    url: '/tags-images/transportation.svg'
   },
   'people': {
     name: 'people',
-    url: '/tags/people.svg'
+    url: '/tags-images/people.svg'
   },
   'parks': {
     name: 'parks',
-    url: '/tags/parks.svg'
+    url: '/tags-images/parks.svg'
   },
   'agora': {
     name: 'agora',
-    url: '/tags/agora.svg'
+    url: '/tags-images/agora.svg'
   },
   'internet': {
     name: 'internet',
-    url: '/tags/internet.svg'
+    url: '/tags-images/internet.svg'
   }
 }
